Refact: "Allow tournament score setting; prevent stage deletion if set exists"
Removed validation that blocked setting score for a tournament match in MatchController. This change accommodates the logical need to set scores for tournament matches. Additionally, in TournamentController, a condition has been introduced to prevent deletion of the round robin phase if there are any matches with sets. This ensures the integrity and simplicity of the tournament management process.

+++ Server/Controllers/TournamentController.cs
@@ -386,11 +386,17 @@ public class TournamentController : ControllerBase
             .Where(x => x.Id == tournamentId)
             .Include(x => x.Creator)
             .Include(x => x.RoundRobinMatches)
+                .ThenInclude(x => x.Sets)
             .SingleAsync(cancellationToken: cancellationToken);
         
         if (tournament.Creator.UserId != userId)
         {
             return Conflict("You are not the creator of this tournament");
+        }
+
+        if (tournament.RoundRobinMatches.Any(x => x.Sets.Any()))
+        {
+            return Conflict("You can't delete the round robin phase if there are matches with sets");
         }
         
         _context.Match.RemoveRange(tournament.RoundRobinMatches);
